+++ spec/rubocop/cop/metrics/block_nesting_spec.rb
@@ -150,7 +150,7 @@ describe RuboCop::Cop::Metrics::BlockNesting, :config do
     expect(cop.offenses.map(&:line)).to eq(lines)
     expect(cop.messages).to eq(
       ['Avoid more than 2 levels of block nesting.'] * lines.length)
-    return unless cop.offenses.size > 0
+    return if cop.offenses.empty?
 
     expect(cop.config_to_allow_offenses['Max']).to eq(max_to_allow)
   end